backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.20
openpyxl==3.1.5
pillow==11.0.0
pytesseract==0.3.13
python-dateutil==2.9.0
